Use WeakHashMap instead of WeakReference
- Fixes .hashCode being different due to a new WeakReference being created everytime an entry is put into cache, as WeakHashMap now handles that automatically for us.
- Fixes WeakReference wrapper object still being leaked even though the object it holds reference into is not. WeakHashMap automatically discards entries of discarded (Garbage Collected) references for us.

Also made a very slight change to fix an instance where NEU was saving a local variable redirect (assigned to value of node.redirect) and doing a null check on it but not using the local variable and instead using node.redirect. Changed it to use the null-checked local variable.

Kept the Collections.synchronizedMap used in LRUCache in the old code for thread safety (wraps the WeakHashMap with the .sychronizedMap), but I do not know if that's needed since MC is single threaded, but I'd suppose its best to not change behaviour of thread-safety.